perf: use u16 for MakeRecord register fields
Change start_reg, count, and dest_reg fields in MakeRecord instruction
from usize to u16 to reduce the Insn enum size. Register indices and
counts are well within u16 range.

+++ core/vdbe/insn.rs
@@ -508,9 +508,9 @@ pub enum Insn {
 
     // Make a record and write it to destination register.
     MakeRecord {
-        start_reg: usize, // P1
-        count: usize,     // P2
-        dest_reg: usize,  // P3
+        start_reg: u16, // P1
+        count: u16,     // P2
+        dest_reg: u16,  // P3
         index_name: Option<String>,
         affinity_str: Option<String>,
     },
